+++ src/hypofuzz/hy.py
@@ -30,6 +30,11 @@ from sortedcontainers import SortedKeyList
 from .corpus import BlackBoxMutator, CrossOverMutator, EngineStub, HowGenerated, Pool
 from .cov import CustomCollectionContext
 
+try:
+    from .debugger import record_pytrace
+except ImportError:
+    record_pytrace = None  # type: ignore
+
 Report = Dict[str, Union[int, float, str, list, Dict[str, int]]]
 
 UNDELIVERED_REPORTS: List[Report] = []
@@ -191,13 +196,20 @@ class FuzzProcess:
         if result.status is Status.INTERESTING:
             # Shrink to our minimal failing example, since we'll stop after this.
             self.shrinking = True
-            Shrinker(
+            shrinker = Shrinker(
                 EngineStub(self._run_test_on, self.random),
                 result,
                 predicate=lambda d: d.status is Status.INTERESTING,
                 allow_transition=None,
-            ).shrink()
+            )
+            shrinker.shrink()
             self.shrinking = False
+            if record_pytrace:
+                # Replay minimal example under our time-travelling debug tracer
+                self._run_test_on(
+                    shrinker.shrink_target.buffer,
+                    collector=record_pytrace(self.nodeid),
+                )
             UNDELIVERED_REPORTS.append(self._json_description)
             self._report_change(UNDELIVERED_REPORTS)
             del UNDELIVERED_REPORTS[:]
@@ -212,7 +224,11 @@ class FuzzProcess:
         #     self.pool.distill(self._run_test_on, self.random)
 
     def _run_test_on(
-        self, buffer: bytes, source: HowGenerated = HowGenerated.shrinking
+        self,
+        buffer: bytes,
+        *,
+        source: HowGenerated = HowGenerated.shrinking,
+        collector: contextlib.AbstractContextManager = None,
     ) -> ConjectureData:
         """Run the test_fn on a given buffer of bytes, in a way a Shrinker can handle.
 
@@ -221,7 +237,8 @@ class FuzzProcess:
         """
         start = time.perf_counter()
         self.ninputs += 1
-        collector = CustomCollectionContext()
+        collector = collector or CustomCollectionContext()  # type: ignore
+        assert collector is not None
         reports: List[str] = []
         argstrings: List[str] = []
         data = ConjectureData(max_length=BUFFER_SIZE, prefix=buffer, random=self.random)
@@ -261,8 +278,10 @@ class FuzzProcess:
         # the `hypothesis.event()` function - exploiting user-defined partitions
         # designed for diagnostic output to guide generation.  See
         # https://hypothesis.readthedocs.io/en/latest/details.html#hypothesis.event
-        data.extra_information.branches = frozenset(collector.branches).union(
-            event_str  # type: ignore  # events aren't branches, but that's OK
+        data.extra_information.branches = frozenset(
+            getattr(collector, "branches", ())  # might be a debug tracer instead
+        ).union(
+            event_str
             for event_str in map(str, data.events)
             if not event_str.startswith("Retried draw from ")
             or event_str.startswith("Aborted test because unable to satisfy ")